Converted api migrations tests to mocha
